requirements.txt
fastapi
uvicorn
pillow
opencv-python
python-multipart
ultralytics
opencv-python-headless